requirements.txt
pyserial==3.5
numpy==1.25.2
scipy==1.11.3
matplotlib==3.8.0